web/requirements.txt
Flask
Flask-Session
wtforms
gunicorn
python-dotenv
redis
bcrypt
PyJWT
requests